+++ laktory/resources/quickstart-stacks/workflows/workspacefiles/notebooks/dlt_laktory_pl.py
@@ -47,37 +47,37 @@ def define_table(node, sink):
     else:
         kwargs["table_properties"] = sink.table_properties
 
-    @table_or_view(**kwargs)
-    @dlt.expect_all(dlt_warning_expectations)
-    @dlt.expect_all_or_drop(dlt_drop_expectations)
-    @dlt.expect_all_or_fail(dlt_fail_expectations)
-    def get_df():
-        # Execute node
-        node.execute()
-        if sink and sink.is_quarantine:
-            df = node.quarantine_df
-        else:
-            df = node.output_df
+    if not sink.is_cdc:
 
-        # Return
-        return df.to_native()
+        @table_or_view(**kwargs)
+        @dlt.expect_all(dlt_warning_expectations)
+        @dlt.expect_all_or_drop(dlt_drop_expectations)
+        @dlt.expect_all_or_fail(dlt_fail_expectations)
+        def get_df():
+            # Execute node
+            node.execute()
+            if sink and sink.is_quarantine:
+                df = node.quarantine_df
+            else:
+                df = node.output_df
 
+            # Return
+            return df.to_native()
 
-# --------------------------------------------------------------------------- #
-# CDC tables                                                                  #
-# --------------------------------------------------------------------------- #
+    else:
 
+        @dlt.view(name=sink.dlt_pre_merge_name)
+        def get_df():
+            node.execute()
+            return node.output_df.to_native()
 
-def define_cdc_table(node, sink):
-    dlt.create_streaming_table(
-        name=sink.dlt_name,
-        comment=node.comment,
-        table_properties=sink.table_properties,
-    )
+        dlt.create_streaming_table(
+            name=sink.dlt_name,
+            comment=node.comment,
+            table_properties=sink.table_properties,
+        )
 
-    dlt.apply_changes(
-        source=node.source.node.primary_sink.dlt_name, **sink.dlt_apply_changes_kwargs
-    )
+        dlt.apply_changes(**sink.dlt_apply_changes_kwargs)
 
 
 # --------------------------------------------------------------------------- #
@@ -90,7 +90,4 @@ for node in pl.nodes:
         continue
 
     for sink in node.sinks:
-        if sink.is_cdc:
-            define_cdc_table(node, sink)
-        else:
-            define_table(node, sink)
+        define_table(node, sink)